backend/requirements.txt
flask==2.1.1
mysql-connector-python==8.0.28
flask-cors==3.1.1
nltk==3.7
